+++ src/sdllib/misc.h
@@ -106,11 +106,11 @@ extern "C" {
 #endif
 
 // Legacy byte generator state behind Random(); the games seed it.
-extern long RandNumb;
+extern int RandNumb;
 
 void* Build_Fading_Table(const void* palette,
-                         void* dest ABSL_ATTRIBUTE_LIFETIME_BOUND,
-                         long int color, long int frac);
+                         void* dest ABSL_ATTRIBUTE_LIFETIME_BOUND, int color,
+                         int frac);
 
 extern int Clip_Rect(int* x, int* y, int* dw, int* dh, int width, int height);
 extern int Confine_Rect(int* x, int* y, int dw, int dh, int width, int height);
